moved shuffle.js to /lib/utils, updated imports, removed unused await statement
